requirements.txt
pyspark==3.5.0
delta-spark==3.1.0
streamlit==1.33.0
pandas==2.0.3
sentence-transformers==2.7.0
torch==2.1.2
chromadb==0.4.24
ortools==9.8.3296
langchain-google-genai==0.1.0
langchain-community==0.0.31
langchain-core==0.1.33
plotly==5.18.0
scikit-learn==1.3.2
faker==24.4.0
numpy==1.26.4
